requirements.txt
aiohttp==3.8.6
aiosignal==1.3.1
async-timeout==4.0.3
attrs==23.1.0
backcall==0.2.0
certifi==2022.9.24
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.6
distlib==0.3.6
executing==1.2.0
filelock==3.9.0
Flask==2.2.2
Flask-WTF==1.0.1
frozenlist==1.4.0
gunicorn==20.1.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
multidict==6.0.4
openai==0.28.1
pickleshare==0.7.5
Pillow==10.0.1
platformdirs==2.6.2
prompt-toolkit==3.0.36
psutil==5.9.4
pure-eval==0.2.2
Pygments==2.13.0
PyMySQL==1.0.2
python-dotenv==0.21.0
pywin32==305
pyzmq==24.0.1
reportlab==4.0.6
requests==2.28.1
schedule==1.1.0
six==1.16.0
tornado==6.2
tqdm==4.66.1
traitlets==5.7.1
urllib3==1.26.13
wcwidth==0.2.5
Werkzeug==2.2.2
WTForms==3.0.1
yarl==1.9.2
